Move `inflateData` to `breadboard` package.
