+++ packages/breadboard/src/data.ts
@@ -25,7 +25,7 @@ export const asBlob = async (
   part: InlineDataCapabilityPart | StoredDataCapabilityPart
 ) => {
   let url: string;
-  if ("storedData" in part) {
+  if (isStoredData(part)) {
     url = part.storedData.handle;
   } else {
     url = `data:${part.inlineData.mimeType};base64,${part.inlineData.data}`;
@@ -33,4 +33,67 @@ export const asBlob = async (
   const response = await fetch(url);
   const data = await response.blob();
   return data;
+};
+
+const isStoredData = (value: unknown): value is StoredDataCapabilityPart => {
+  if (typeof value !== "object" || value === null) return false;
+  const data = value as DataCapability;
+  if (!("storedData" in data)) return false;
+  if (!data.storedData.handle) return false;
+  return true;
+};
+
+function asBase64(file: File | Blob): Promise<string> {
+  return new Promise((resolve, reject) => {
+    const reader = new FileReader();
+    reader.onload = () => {
+      if (typeof reader.result !== "string") {
+        reject("Reader result is not a string");
+        return;
+      }
+
+      const [, content] = reader.result.split(",");
+      resolve(content);
+    };
+    reader.onerror = (err) => reject(err);
+    reader.readAsDataURL(file);
+  });
+}
+
+/**
+ * Recursively descends into the data object and inflates any
+ * `StoreDataCapabilityPart`, turning it into
+ * `InlineDataCapabilityPart`.
+ * @param data -- data to inflate
+ * @returns -- a new object with all `StoredDataCapabilityPart`
+ * replaced with `InlineDataCapabilityPart`
+ */
+export const inflateData = async (data: unknown) => {
+  const descender = async (value: unknown): Promise<unknown> => {
+    if (isStoredData(value)) {
+      const { mimeType, handle } = value.storedData;
+      const blob = await (await fetch(handle)).blob();
+      const data = await asBase64(blob);
+      return { inlineData: { data, mimeType } };
+    }
+    if (Array.isArray(value)) {
+      const result = [];
+      for (const item of value) {
+        result.push(await descender(item));
+      }
+      return result;
+    }
+    if (typeof value === "object" && value !== null) {
+      const v = value as Record<string, unknown>;
+      const result: Record<string, unknown> = {};
+      for (const key in value) {
+        result[key] = await descender(v[key]);
+      }
+      return result;
+    }
+    return value;
+  };
+
+  const result = await descender(data);
+  return result;
 };
